% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crop.circle.R
\name{crop.circle}
\alias{crop.circle}
\title{crop.circle}
\usage{
crop.circle(x3p.surf.file.info, radius)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
\code{spatstat}
}

