% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crop.retangle.R
\name{crop.retangle}
\alias{crop.retangle}
\title{crop.retangle}
\usage{
crop.retangle(x3p.surf.file.info, percent.off.each.side = 1)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
\code{spatstat}
}

