% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fill.holes.R
\name{fill.holes}
\alias{fill.holes}
\title{fill.holes}
\usage{
fill.holes(x3p.surf.file.info, hole.obj, big.hole.tol = 2,
  big.hole.fill.typ = "surface.mean", hole.window.factor = 1.5,
  fill.type.df = data.frame("all.the.same", "1D.interpolation"),
  printQ = FALSE)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
\code{spatstat}
}

