% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find.holes.R
\name{find.holes}
\alias{find.holes}
\title{find.holes}
\usage{
find.holes(dmat, plotQ = FALSE)
}
\arguments{
\item{dmat}{The surface of points}
}
\value{
A \code{spatstat} pixel image object (object of class im).
}
\description{
XXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
\code{spatstat}
}

