% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/launch_3Dview.R
\name{launch_3Dview}
\alias{launch_3Dview}
\title{Shiny app to view a 3D surface}
\usage{
launch_3Dview()
}
\description{
Shiny app to view a 3D surface
}
\details{
Should work with all supported 3D file formats.
}
\examples{
launch_3Dview()
}
\references{
http://shiny.rstudio.com/
}

