% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lms2x3p.file.R
\name{lms2x3p.file}
\alias{lms2x3p.file}
\title{Convert a Zeiss LMS file to an X3P format.}
\usage{
lms2x3p.file(surf.info, extra.file.info.list, comment.list, fname,
  move.to.directory)
}
\arguments{
\item{surf.info}{surface information returned from read.zeiss.lms.file}

\item{extra.file.info.list}{extra information required by ISO standards}

\item{comment.list}{optional comments on how the data was obtained.}

\item{fname}{name for the .x3p file to be written}

\item{move.to.directory}{path where to move the file.}
}
\value{
Nothing.
}
\description{
Convert a Zeiss LMS file to an X3P format.
}
\details{
Convert a Zeiss LMS file to an X3P format and writes the result to file. \code{move.to.directory}
is the path where to move the file. The file itself is initiall written to \code{tempdir()}.
}
\examples{
Coming soon.

\dontrun{

}
}

