% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.x3p.header.R
\name{make.x3p.header}
\alias{make.x3p.header}
\title{Put together the XML format x3p header following the ISO 25178 and ISO 5436-2 standards.}
\usage{
make.x3p.header(surf.type, x.axis.type, x.data.type, xinc, x.offset,
  y.axis.type, y.data.type, yinc, y.offset, z.axis.type, z.data.type, zinc,
  z.offset, meas.dte, who.wrote.file, manufacturer, model, sn, vers, cal.dte,
  probe.type, probe.id, meas.comment, pts.per.prof, num.prof, z.format,
  data.bin.md5sumcheck)
}
\arguments{
\item{surf.type}{}

\item{x.axis.type}{}

\item{x.data.type}{}

\item{xinc}{}

\item{x.offset}{}

\item{y.axis.type}{}

\item{y.data.type}{}

\item{yinc}{}

\item{y.offset}{}

\item{z.axis.type}{}

\item{z.data.type}{}

\item{zinc}{}

\item{z.offset}{}

\item{meas.dte}{}

\item{who.wrote.file}{}

\item{manufacturer}{}

\item{model}{}

\item{sn}{}

\item{vers}{}

\item{cal.dte}{}

\item{probe.type}{}

\item{probe.id}{}

\item{meas.comment}{}

\item{pts.per.prof}{}

\item{num.prof}{}

\item{z.format}{}

\item{data.bin.md5sumcheck}{}
}
\value{
An object of class \code{XMLNode}
}
\description{
Put together the XML format x3p header, main.xml, following the ISO 25178 and ISO 5436-2 standards.
}
\details{
Put together the XML format x3p header following the ISO 25178 and ISO 5436-2 standards. The
arguements to this function are the obligatory fields that must appear in the x3p header
according to the ISO 5436-2 standard.
}
\references{
http://open-gps.sourceforge.net/
}

