% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot3D.x3p.file.R
\name{plot3D.x3p.file}
\alias{plot3D.x3p.file}
\title{3D plot  of a Digital Surf formated surface}
\usage{
plot3D.x3p.file(x3p.surf.file.info, num.x.pts = NULL, num.slices = NULL,
  aspect = c(1, 0.3, 0.2), plot.type = "points")
}
\arguments{
\item{x3p.surf.file.info}{}

\item{num.x.pts}{}

\item{num.slices}{}

\item{aspect}{}

\item{plot.type}{}
}
\value{
a plot.
}
\description{
3D plot  of a Digital Surf formated surface
}
\details{
none.
}
\examples{
Coming soon.

\dontrun{

}
}
\references{
http://open-gps.sourceforge.net/
}

