% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.x3p.file.info.R
\name{print.x3p.file.info}
\alias{print.x3p.file.info}
\title{Format (most) x3p header file info into a nice
read-able format}
\usage{
\method{print}{x3p.file.info}(main.xml.info)
}
\arguments{
\item{main.xml.info}{}
}
\value{
a string.
}
\description{
Format (most) x3p header file info into a nice
read-able format
}
\details{
none.
}
\examples{
Coming soon.

\dontrun{

}
}
\references{
http://open-gps.sourceforge.net/
}

