% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.digital.surf.file.R
\name{read.digital.surf.file}
\alias{read.digital.surf.file}
\title{Read in a Digital Surf file wrapper. Just specify the path to the
.sur file.}
\usage{
read.digital.surf.file(file.path)
}
\arguments{
\item{file.path}{}
}
\value{
a matrix.
}
\description{
Read in a Digital Surf file wrapper. Just specify the path to the
.sur file.
}
\details{
Use fle.choose() to open a file chooser window
}
\examples{
Coming soon.

\dontrun{

}
}
\references{
http://open-gps.sourceforge.net/
}

