% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.digital.surf.file2.R
\name{read.digital.surf.file2}
\alias{read.digital.surf.file2}
\title{Faster read in a Digital Surf file wrapper. Calls a faster C++ routine
via Rcpp.}
\usage{
read.digital.surf.file2(file.path)
}
\arguments{
\item{file.path}{}
}
\value{
an array.
}
\description{
Faster read in a Digital Surf file wrapper. Calls a faster C++ routine
via Rcpp.
}
\details{
Just specify the path to the
.sur file. Use fle.choose() to open a file chooser window
}
\examples{
Coming soon.

\dontrun{

}
}
\references{
http://open-gps.sourceforge.net/
}

