% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.digital.surf.header.R
\name{read.digital.surf.header}
\alias{read.digital.surf.header}
\title{pt = pointer returned by calling file on a digital surf .sur or
.pro binary format file}
\usage{
read.digital.surf.header(pt)
}
\arguments{
\item{pt}{}
}
\value{
a vector.
}
\description{
pt = pointer returned by calling file on a digital surf .sur or
.pro binary format file
}
\details{
**SEE FORMATS SECTION OF HELP IN MOUNTAINS SOFTWARE FOR
INFORMATION ON THE .sur and .pro formats
}
\examples{
Coming soon.

\dontrun{

}
}
\references{
http://open-gps.sourceforge.net/
}

