% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.digital.surf.profiles.R
\name{read.digital.surf.profiles}
\alias{read.digital.surf.profiles}
\title{Read in the profiles once the header info is known
Fastest version}
\usage{
read.digital.surf.profiles(pt, header.info)
}
\arguments{
\item{pt}{}

\item{header.info}{}
}
\value{
not sure.
}
\description{
Read in the profiles once the header info is known
Fastest version
}
\details{
none.
}
\examples{
Coming soon.

\dontrun{

}
}
\references{
http://open-gps.sourceforge.net/
}

