% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.x3p.R
\name{read.x3p}
\alias{read.x3p}
\title{Read an x3p file}
\usage{
read.x3p(fpath)
}
\arguments{
\item{fpath}{}
}
\value{
a list
}
\description{
Read an x3p file
}
\details{
Unzips file to tempdir().Returns a list. The first element is the x3p file header. The
second element is the surface data in the form of a matrix.
}
\examples{
file.path <- system.file("extdata", "glock.x3p", package="x3pr")
glock.x3p.info <- read.x3p(file.path)
glock.x3p.info[[1]]
}
\references{
http://open-gps.sourceforge.net/
}

