% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.zeiss.lms.file.R
\name{read.zeiss.lms.file}
\alias{read.zeiss.lms.file}
\title{Read a surface file in Zeiss LMS format.}
\usage{
read.zeiss.lms.file(file.path)
}
\arguments{
\item{file.path}{}
}
\value{
a list. The Zeiss LMS header info and the surface.
}
\description{
Read a surface file in Zeiss LMS format.
}
\details{
Read a surface file in Zeiss LMS format.
}
\examples{
Coming soon.

\dontrun{

}
}

