% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.zeiss.lms.header.R
\name{read.zeiss.lms.header}
\alias{read.zeiss.lms.header}
\title{Read a header of surface file in Zeiss LMS format.}
\usage{
read.zeiss.lms.header(pt)
}
\arguments{
\item{pt}{R file pointer to the surface file}
}
\value{
NOTING
}
\description{
Read header ofsurface file in Zeiss LMS format.
}
\details{
Read header of surface file in Zeiss LMS format.
}
\examples{
Coming soon.

\dontrun{

}
}

