% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.zeiss.lms.profiles.R
\name{read.zeiss.lms.profiles}
\alias{read.zeiss.lms.profiles}
\title{Read out surface data in Zeiss LMS format.}
\usage{
read.zeiss.lms.profiles(pt, header.info)
}
\arguments{
\item{pt}{a pointer to a Zeiss LMS file}
}
\value{
a matrix. The surface data in the Zeiss LMS file.
}
\description{
Read out surface data in Zeiss LMS format.
}
\details{
Read a surface file in Zeiss LMS format. This function is
formulated as a helper to read.zeiss.lms.file and refered to as ~.profiles
for historical reasons.
}
\examples{
Coming soon.

\dontrun{

}
}
\seealso{
read.zeiss.lms.file
}

