% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sur2x3p.file.R
\name{sur2x3p.file}
\alias{sur2x3p.file}
\title{Convert a read in .sur file to x3p format}
\usage{
sur2x3p.file(surf.info, extra.file.info.list, comment.list, fname,
  move.to.directory)
}
\arguments{
\item{surf.info}{}

\item{extra.file.info.list}{}

\item{comment.list}{}

\item{fname}{}

\item{move.to.directory}{}
}
\value{
a x3p file.
}
\description{
Convert a read in .sur file to x3p format
}
\details{
It's necessary at this point to open the .sur file so
that point's can be converted to floats (or doubles),
which as far as I can tell is required by the
NanoFicus x3p library
}
\examples{
Coming soon.

\dontrun{

}
}
\references{
http://open-gps.sourceforge.net/
}

