% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.x3p.R
\name{write.x3p}
\alias{write.x3p}
\title{Write a x3p format file.}
\usage{
write.x3p(x3p.header.info.list, surface.mat, fname, move.to.directory)
}
\arguments{
\item{x3p.header.info.list}{XXXXXXXXXXX}

\item{surface.mat}{XXXXXXXXXXX}

\item{fname}{desired .x3p file name}

\item{move.to.directory}{desired place to put the .x3p file}
}
\value{
a x3p file is written to disk.
}
\description{
Write a x3p format file.
}
\details{
Write a x3p format file.
}
\examples{
Comming soon.
}
\references{
http://open-gps.sourceforge.net/
}

