\name{x3pr-package}
\alias{x3pr-package}
\alias{x3pr}
\docType{package}
\title{
Read/Write functionality for X3P surface metrology format
}
\description{
R Read/Write and interactive 3D view functionality for standardized surface metrology format X3P. Cf.: \cr
http://www.opengps.eu/2008/ISO5436_2\cr 
http://www.opengps.eu/2008/ISO5436_2/ISO5436_2.xsd\cr
\cr
The package also has functionality to read Digital-Surf (ver. 5) format and Zeiss LMS format.
}
\details{
\tabular{ll}{
Package: \tab x3pr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-11\cr
License: \tab GPL 3\cr
}
R Read/Write functionality for standardized surface metrology format X3P. \code{read.x3p(file.path)}
}
\author{
OpenFMC group: https://groups.google.com/forum/#!forum/openfmc \cr
Maintainer: Nick Petraco <npetraco@gmail.com>
}
\references{
None yet
}
Surface
Metrology
3D
\keyword{ package }
\examples{
%% ~~ simple examples of the most important functions ~~
}
