% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot3D.x3p.file.R
\name{plot3D.x3p.file}
\alias{plot3D.x3p.file}
\title{3D plot  of a surface in X3P format}
\usage{
plot3D.x3p.file(x3p.surf.file.info, num.x.pts = NULL, num.slices = NULL,
  aspect = c(1, 0.3, 0.2), plot.type = "points", color = if (plot.type ==
  "points") "black" else rev(rainbow(256, start = 0/6, end = 2/6)))
}
\arguments{
\item{x3p.surf.file.info}{information read in by \code{read.x3p}}

\item{num.x.pts}{number of points desired in the x-direction.}

\item{num.slices}{number of points desired in the y-direction.}

\item{aspect}{aspect ratios}

\item{plot.type}{"points" or "surface"}

\item{color}{The color(s) to use for the plot}
}
\value{
An RGL 3D plot.
}
\description{
3D plot  of a surface in X3P format
}
\details{
3D plot  of a surface in X3P format. Some type of decimation is usually
required so that the surface can be rendered fast. The arguments num.x.pts
and num.slices allow the user to subsample the surface down to the desired size.
If the number of points is still potentially large, the function will prompt the
user if they want to continue.
}
\examples{
file.path <- system.file("extdata", "glock.x3p", package="x3pr")
glock.x3p.info <- read.x3p(file.path)
plot3D.x3p.file(glock.x3p.info, 1024, 80, aspect=c(1,3,0.4), plot.type="surface")
plot3D.x3p.file(glock.x3p.info, 1024, 80, aspect=c(1,3,0.4), plot.type="surface", color = topo.colors(256))
}
\references{
http://open-gps.sourceforge.net/
}

