% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.x3p.file.info.R
\name{print.x3p.file.info}
\alias{print.x3p.file.info}
\title{Format (most) x3p header file info into a nice
read-able format}
\usage{
\method{print}{x3p.file.info}(main.xml.info)
}
\arguments{
\item{main.xml.info}{}
}
\value{
a string.
}
\description{
Format (most) x3p header file info into a nice
read-able format
}
\details{
none.
}
\examples{
file.path <- system.file("extdata", "glock.x3p", package="x3pr")
glock.x3p.info <- read.x3p(file.path)
glock.x3p.info[[1]]
print.x3p.file.info(glock.sur.surface.info[[1]]z)
}
\references{
http://open-gps.sourceforge.net/
}

