% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.digital.surf.file.R
\name{read.digital.surf.file}
\alias{read.digital.surf.file}
\title{Read in a Digital Surf file.}
\usage{
read.digital.surf.file(file.path)
}
\arguments{
\item{file.path}{}
}
\value{
a matrix.
}
\description{
Read in a Digital Surf file. Just specify the path to the
.sur file.
}
\details{
This function calls \code{read.digital.surf.header} and \code{read.digital.surf.profiles}.
}
\examples{
file.path <- system.file("extdata", "glock.sur", package="x3pr")
glock.sur.info <- read.digital.surf.file(file.path)
glock.sur.info[[1]]
}
\references{
http://open-gps.sourceforge.net/
}

