% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.digital.surf.header.R
\name{read.digital.surf.header}
\alias{read.digital.surf.header}
\title{Read the header off a Digital Surf file.}
\usage{
read.digital.surf.header(pt)
}
\arguments{
\item{pt}{pointer returned by calling \code{read.digital.surf.file}}
}
\value{
a list.
}
\description{
pt = pointer returned by calling file on a digital surf .sur or
.pro binary format file
}
\details{
**SEE FORMATS SECTION OF HELP IN MOUNTAINS SOFTWARE FOR
INFORMATION ON THE .sur and .pro formats. Also see code for \code{read.digital.surf.file}.
}
\references{
http://open-gps.sourceforge.net/
}

