% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.digital.surf.profiles.R
\name{read.digital.surf.profiles}
\alias{read.digital.surf.profiles}
\title{Read in the surface data of a Digital Surf file once the header info is known.}
\usage{
read.digital.surf.profiles(pt, header.info)
}
\arguments{
\item{pt}{}

\item{header.info}{}
}
\value{
not sure.
}
\description{
Read in the surface data of a Digital Surf file once the header info is known.
}
\details{
Called by \code{read.digital.surf.file}.
}
\references{
http://open-gps.sourceforge.net/
}

